Add files via upload
A multilayer light reflection model is designed based on the stacking unit of oxide films, metallic film, and vacuum gap. By tuning the number of units, the thickness of each layer, and the stacking sequence, an effective light-reflection medium is expected to be achieved on the building windows or vehicle windows. Beyond the model design, the genetic algorithm is used as an optimization method to optimize these variables. In the python-based simulation, an initial demonstration is introduced to pave a direction for maximizing the light reflectance. 
The OptSimulations.py is defined to modify the model stacking sequence and other parameters. With the accompany of ds.py and plot_r_a.py, the model-generated reflectance, absorption, transmission, and extinction data is handled and plotted. The reflection calculation core is based on the tmm_core.py, referred to as the contribution of Steve Byrnes (https://arxiv.org/abs/1603.02720).
Using the genetic algorithm, the variables such as precision, population size, and generations are the key parameters to optimize the model. In order to keep the transparency of the film layers, the variable variation range determines the optimizing computation cost as well as the manufacturing difficulty. The genetic algorithm-based optimization is generated in the optical wavelength range of 400nm to 800 nm. In real and complete modeling, a wider range is also needed.